github-122: take scrubbing invocation out of payload and move it to RollbarLogger
