+++ src/Config.php
@@ -334,9 +334,9 @@ class Config
         return error_reporting() === 0 && !$this->reportSuppressed;
     }
 
-    public function send($payload, $accessToken)
+    public function send(&$scrubbedPayload, $accessToken)
     {
-        return $this->sender->send($payload, $accessToken);
+        return $this->sender->send($scrubbedPayload, $accessToken);
     }
 
     public function handleResponse($payload, $response)
